Unify unittests for WSIReader (#3261)
* Unify wsireader unittests and add new tests


* Remove variant of wsireader unittests


* Fix import


* Change the logic to skip


* Ignore N802 for setUpModule


* Skip if no tifffile


* exclude test_wsireader form min tests


* Remove trailing comma


* Change WSIReaderTests base class
