+++ tests/min_tests.py
@@ -104,7 +104,6 @@ def run_testsuit():
         "test_nifti_rw",
         "test_nifti_saver",
         "test_occlusion_sensitivity",
-        "test_openslide_reader",
         "test_orientation",
         "test_orientationd",
         "test_parallel_execution",
@@ -140,6 +139,7 @@ def run_testsuit():
         "test_vit",
         "test_vitautoenc",
         "test_write_metrics_reports",
+        "test_wsireader",
         "test_zoom",
         "test_zoom_affine",
         "test_zoomd",
